requirements.txt
networkx
pyyaml
fire
pytest